Bug Fix - DynamirFrame version 2 was not detected during decoding
